refactor(processor): split processor.rs into modular submodules
Split the 1160-line processor.rs into a processor/ module directory:
- state.rs: RuntimeModSources, RuntimeVoiceEnv/Line/DebugState, NoteEntry,
  MonoStackEntry, CzDacColor
- notes.rs: note-on/off, sustain, voice allocation and management
- process.rs: audio process loop with mod source computation and rendering
- utils.rs: midi_note_to_freq, soft_clip_tanh, signed_pow, one-pole filters
- mod.rs: CosmoProcessor struct, new(), FX/param/MIDI methods, tests

All 43 tests pass (2 pre-existing failures unrelated to this change).
